+++ executor/host/deposit.go
@@ -16,7 +16,7 @@ func (h *Host) initiateDepositHandler(_ context.Context, args nodetypes.EventHan
 	if err != nil {
 		return err
 	}
-	if bridgeId != uint64(h.BridgeId()) {
+	if bridgeId != h.BridgeId() {
 		// pass other bridge deposit event
 		return nil
 	}
